TASK-4017 Overwrite non-config resources on each plugin startup
We just wondered why a server still had an outdated mcpi library; the reason is that once the files have been written to the file system (in fact, even once the directory those are in exists), all those resources aren't written.
This makes sense for the plugin configuration, but cmdsvr, mcpi, and pplugins are all delivered by JuicyRaspberryPie and should not be modified by the user.
Don't skip saveResource() when the directory already exists and set the overwrite flag, too. This also reduces the complexity of the code a bit.