TASK-4337 Add plugin classloader
To be able to access more than just standard Java classes.
Note that as these are still loaded via a custom JShell classloader, these are separate instances, and JavaPlugin objects cannot be instantiated, because they miss the connection to the original server's classloader. We'd have to inject the actual, existing plugin objects for that to work.